+++ packages/@tinacms/core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.4](https://github.com/tinacms/tinacms/compare/@tinacms/core@0.7.3...@tinacms/core@0.7.4) (2020-04-06)
+
+**Note:** Version bump only for package @tinacms/core
+
+
+
+
+
 ## [0.7.3](https://github.com/tinacms/tinacms/compare/@tinacms/core@0.7.2...@tinacms/core@0.7.3) (2020-03-16)
 
 **Note:** Version bump only for package @tinacms/core