+++ packages/gatsby-tinacms-git/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.7-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-git@0.4.6...gatsby-tinacms-git@0.4.7-alpha.0) (2020-04-06)
+
+**Note:** Version bump only for package gatsby-tinacms-git
+
+
+
+
+
+## [0.4.6](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-git@0.4.5...gatsby-tinacms-git@0.4.6) (2020-04-06)
+
+**Note:** Version bump only for package gatsby-tinacms-git
+
+
+
+
+
 ## [0.4.5](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-git@0.4.4...gatsby-tinacms-git@0.4.5) (2020-03-30)
 
 **Note:** Version bump only for package gatsby-tinacms-git