+++ packages/gatsby-tinacms-json/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.18-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.17...gatsby-tinacms-json@0.8.18-alpha.0) (2020-04-06)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
+## [0.8.17](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.16...gatsby-tinacms-json@0.8.17) (2020-04-06)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
 ## [0.8.16](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.15...gatsby-tinacms-json@0.8.16) (2020-03-30)
 
 **Note:** Version bump only for package gatsby-tinacms-json