+++ packages/next-tinacms-json/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.11-alpha.0](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.10...next-tinacms-json@0.3.11-alpha.0) (2020-04-06)
+
+**Note:** Version bump only for package next-tinacms-json
+
+
+
+
+
+## [0.3.10](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.9...next-tinacms-json@0.3.10) (2020-04-06)
+
+**Note:** Version bump only for package next-tinacms-json
+
+
+
+
+
 ## [0.3.9](https://github.com/tinacms/tinacms/compare/next-tinacms-json@0.3.8...next-tinacms-json@0.3.9) (2020-03-30)
 
 **Note:** Version bump only for package next-tinacms-json