+++ packages/react-tinacms/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.19-alpha.0](https://github.com/tinacms/tinacms/compare/react-tinacms@0.9.18...react-tinacms@0.9.19-alpha.0) (2020-04-06)
+
+**Note:** Version bump only for package react-tinacms
+
+
+
+
+
+## [0.9.18](https://github.com/tinacms/tinacms/compare/react-tinacms@0.9.17...react-tinacms@0.9.18) (2020-04-06)
+
+**Note:** Version bump only for package react-tinacms
+
+
+
+
+
 ## [0.9.17](https://github.com/tinacms/tinacms/compare/react-tinacms@0.9.16...react-tinacms@0.9.17) (2020-03-30)
 
 **Note:** Version bump only for package react-tinacms